+++ contracts/test/TestExecutor.sol
@@ -4,23 +4,19 @@ pragma solidity 0.6.12;
 import "../interfaces/IExecutor.sol";
 
 contract TestExecutor is IExecutor {
-
     event Executed(
-        uint indexed _id,
-        uint _for,
-        uint _against,
-        uint _quorum
+        uint256 indexed _id,
+        uint256 _for,
+        uint256 _against,
+        uint256 _quorum
     );
 
     function execute(
-        uint _id,
-        uint _for,
-        uint _against,
-        uint _quorum
-    )
-    external
-    override
-    {
+        uint256 _id,
+        uint256 _for,
+        uint256 _against,
+        uint256 _quorum
+    ) external override {
         emit Executed(_id, _for, _against, _quorum);
     }
 }
